Direct connect and online lobby now try other ports
The server can manually port forward his port and when the client direct
connect or use online lobby the engine will now try different ports to
connect to the server. A port that got portforward can also be provided
from the server to the client via lobby. Check Tip in code.