Punch will now try sequence ports and try predict the ports
For NATs that change the external port.
Clients save the port received from the master server to the game
server. If the game server restart the client will get the new port and
try to predict the port the game server use. The client will also try
ports near the received port. To test more ports increase the connection
timeout.