Fix unsync bug
ds_map_size(m) is changed when use ds_map_delete(m,insthash); and the
loop will quit early. Before thw whole map is checked. We must preserve
the m map to the last end or redo the loop and set i=0. Tested with 5
players and all works well with this fix.