Delta time to connection
Connection will use delta time to allow the server and clients to have
different room speeds. Else the say the server got room speed 60 and the
client 30. Then the server will kick the client because the client ping
timeout is only half and the servers ping timeout will run out and kick
the client.

NOTE:
The  mp_sync still use the same step counter as before.