Error when connection fail
When connection fail the client is still in the connecting room. That is
because the script udphp_stopClient() destroy the obj_htme. And the
script htme_clientConnectionFailed dont check if the obj_htme still
exists. The scripts htme_clientIsConnected and htme_debugOverlayEnabled
is also fixed to handle this new client stop procedure.