obj punch now destroyed
Fixed so the obj punch is now destroyed. The clean script reseted the
udphp_client_id to 0 and 0 in the demo is the obj_htme. When the
obj_htme is destroyed the clean script is run again and caused errors.
Fixed so the clean script can run several times without cause errors.