Check if all variables got here
When using SMART some variables will not be sent to the other end. But
the line:
var length = buffer_read(in_buff,buffer_u8);
get the length of the total variables count. But if the varaible is the
same value or in tolerance it won't get here. So we may read outside the
buffer. In gm 1.4.1567 this is ok. It return -3. But in 1.4.1749 GM now
always check if we read outside the buffer and shoot us. This will make
us bulletproof :-)